Add fair unlocking mechanism to RawMutexImpl
Added an implementation of the RawMutexFair trait for the RawMutexImpl struct to ensure fairer unlocking. This consists of a new constant FAIR_BIT and the new function unlock_fair. The modifications ensure that when a waiter is notified, the FAIR_BIT state is also set, minimizing chances of thread starvation.