Add timed locks to Mutex implementation
The update adds support for timed locks to the RawMutex and Mutex structs. This is done through the implementation of the RawMutexTimed trait for RawMutex, which adds try_lock_for and try_lock_until methods. Corresponding changes were made in the Mutex struct, with similar try_lock_for and try_lock_until methods introduced.

This change improves the usage flexibility of the Mutex by providing the option of attempting to acquire the lock for a limited duration. Additional test cases were also added to verify the new functionality.