Add num_waiting method and improve thread sleep
A new method, num_waiting, was added to provide information about the number of waiting threads in the event structure. The portions of the code related to putting threads to sleep has been modified to use this method, providing a more safe and accurate way of handling thread synchronization. Additionally, an assert in mini_lock has been changed to a debug_assert to reduce potential performance impact in release builds.