Add must_use attribute and warning for undocumented unsafe blocks
Added #[must_use] attributes in several places across the code for constructors and other methods that return some kind of value, to prevent unaware loss of this value by not binding it. Also added clippy lint rule for undocumented unsafe blocks to catch potentially dangerous, undocumented unsafe operations in the code which can be hazardous when combined with other commands.